Allow open source licenses to be packaged
A number of open source licenses require that a copy of that license is
distributed alongside the code that uses it, which is currently fairly
annoying to do.

Initially I put only `"LICENSE.txt"` in the list, however very
inconveniently a number of different licenses (eg LGPL, Apache) require
*two* files to be distributed under a variety of names.

To that end, I've added support for a `license` folder in the root of
the addon which will be ignored by the game but will allow authors to
distribute an arbitrary number of text files to freely express whatever
complex licenses or sublicenses apply to the distributed code and content.